requirements.txt
huggingface-hub
llama-cpp-python==0.1.4